Refactored and split the game page.tsx into different parts to make it more modular and added the Itemshop UI, all the coins and item effects logic and the effects themselves (#98, #99)
