+++ app/games/[gameSessionId]/page.tsx
@@ -1,235 +1,90 @@
 "use client";
 
-import { useRouter, useParams } from "next/navigation";
+import { useEffect, useState } from "react";
+import { useParams } from "next/navigation";
 import CodosseumLogo from "@/components/CodosseumLogo";
 import styles from "@/styles/game.module.css";
-import resultStyles from "@/styles/results.module.css";
-import { useEffect, useRef, useState } from "react";
 import useLocalStorage from "@/hooks/useLocalStorage";
 import { getApiDomain } from "@/utils/domain";
-import {
-  SendOutlined,
-  TrophyOutlined,
-  ClockCircleOutlined,
-  PlayCircleOutlined,
-  LoadingOutlined,
-  CheckCircleOutlined,
-} from "@ant-design/icons";
-import CodeMirror from "@uiw/react-codemirror";
-import { indentUnit } from "@codemirror/language";
-import { python } from "@codemirror/lang-python";
-import { java } from "@codemirror/lang-java";
-import SockJS from "sockjs-client";
-import { Client, IMessage } from "@stomp/stompjs";
-import CodosseumAvatar from "@/components/CodosseumAvatar";
-
-interface GameRoundData {
-  gameSessionId: number;
-  playerSessionId: number;
-  playerId: number;
-  currentScore: number;
-  numOfSkippedProblems: number;
-  problemId: number;
-  title: string;
-  description: string;
-  inputFormat: string;
-  outputFormat: string;
-  constraints: string;
-  gameLanguage: string;
-  opponentName?: string;
-  playerAvatarId?: number;
-  opponentAvatarId?: number;
-  endsAt?: string;
-  serverTime?: string;
-}
-
-interface RunTestCase {
-  testCaseId: number;
-  expectedOutput: string;
-  actualOutput: string;
-  result: "PASS" | "FAIL";
-  errorMessage: string | null;
-}
-
-interface ExecutionResult {
-  message?: string;
-  status: "success" | "error" | "info";
-  testCases?: RunTestCase[];
-  summary?: string;
-}
-
-interface GameSessionSampleSolutionsDTO {
-  problemTitle: string;
-  problemSampleSolution: string;
-}
-
-interface PlayerScoreDTO {
-  playerSessionId: number;
-  userId: number;
-  username: string;
-  score: number;
-}
-
-interface GameEndDTO {
-  gameSessionId: number;
-  gameStatus: string;
-  gameEndReason: string;
-  winnerPlayerId: number;
-  playerScores: PlayerScoreDTO[];
-  gameSessionSampleSolutions: Record<string, GameSessionSampleSolutionsDTO>;
-}
-
-interface PlayerGameSummaryDTO {
-  playerSessionId: number;
-  playerId: number;
-  problemResults: {
-    solvedCorrectly: number[];
-    notSolvedFullyCorrectly: number[];
-  };
-}
-
-interface Particle {
-  id: number;
-  x: number;
-  y: number;
-  size: number;
-  color: string;
-  duration: number;
-}
+import { CheckCircleOutlined } from "@ant-design/icons";
+import { useGameWebSocket } from "./_hooks/useGameWebSocket";
+import { useGameTimer } from "./_hooks/useGameTimer";
+import { useScoreAnimation } from "./_hooks/useScoreAnimation";
+import GameOverScreen from "./_components/GameOverScreen";
+import ScoreBox from "./_components/ScoreBox";
+import ProblemPanel from "./_components/ProblemPanel";
+import CodeEditorPanel from "./_components/CodeEditorPanel";
+import ItemShop from "./_components/ItemShop";
+import { ExecutionResult, GameRoundData, Problem } from "./_types";
 
 const GAME_DURATION_MS = 15 * 60 * 1000;
-const PARTICLE_COLORS = ["#eab308", "#fbbf24", "#f59e0b", "#ffffff", "#fde68a", "#facc15"];
 
-function formatTime(seconds: number): string {
-  const m = Math.floor(seconds / 60);
-  const s = seconds % 60;
-  return `${m}:${s.toString().padStart(2, "0")}`;
-}
-
-function generateParticles(): Particle[] {
-  return Array.from({ length: 18 }, (_, i) => {
-    const angle = (i / 18) * 360 + (Math.random() - 0.5) * 25;
-    const distance = 80 + Math.random() * 80;
-    const rad = (angle * Math.PI) / 180;
-    return {
-      id: Date.now() + i,
-      x: Math.cos(rad) * distance,
-      y: Math.sin(rad) * distance,
-      size: 8 + Math.random() * 9,
-      color: PARTICLE_COLORS[Math.floor(Math.random() * PARTICLE_COLORS.length)],
-      duration: 0.9 + Math.random() * 0.3,
-    };
-  });
-}
-
-export default function GamePage() {
-  const router = useRouter();
-  const { value: userId } = useLocalStorage("userid", "");
-  const { value: token } = useLocalStorage("token", "");
-  const { value: storedUsername } = useLocalStorage("username", "Player One");
-  const { value: storedAvatarId } = useLocalStorage("avatarId", "1");
-  const { gameSessionId } = useParams();
-
-  const [language, setLanguage] = useState("python");
-  const [playerSessionId, setPlayerSessionId] = useState<number | null>(null);
-  const [players, setPlayers] = useState<Record<string, { username: string; score: number }>>({});
-
-  const me = players[String(userId)];
-  const allPlayers = Object.entries(players);
-
-  const opponentEntry = userId
-    ? allPlayers.find((entry): entry is [string, { username: string; score: number }] => {
-        const [id] = entry;
-        return id !== String(userId);
-      })
-    : null;
-  const opponent = opponentEntry ? opponentEntry[1] : null;
-
-  const myScore = me?.score ?? 0;
-
-  const [playerAvatarId, setPlayerAvatarId] = useState<number>(1);
-  const [opponentAvatarId, setOpponentAvatarId] = useState<number>(2);
-
-  const [problem, setProblem] = useState<{
-    id: number;
-    title: string;
-    description: string;
-    inputFormat: string;
-    outputFormat: string;
-    constraints: string;
-  } | null>(null);
-
-  const [currentRound, setCurrentRound] = useState(1);
-  const [isGameOver, setIsGameOver] = useState(false);
-  const [gameEndData, setGameEndData] = useState<GameEndDTO | null>(null);
-  const [gameSummary, setGameSummary] = useState<PlayerGameSummaryDTO | null>(null);
-  const [expandedSolutions, setExpandedSolutions] = useState<Set<string>>(new Set());
-
-  // Timer
-  const [gameEndTime, setGameEndTime] = useState<number | null>(null);
-  const [timeLeft, setTimeLeft] = useState<number | null>(null);
-
-  // Submit toast
-  const [showSubmitToast, setShowSubmitToast] = useState(false);
-
-  // Score animations
-  const prevMyScore = useRef(0);
-  const prevOpponentScore = useRef(0);
-  const [myFlash, setMyFlash] = useState<{ delta: number; key: number } | null>(null);
-  const [opponentFlash, setOpponentFlash] = useState<{ delta: number; key: number } | null>(null);
-  const [myParticles, setMyParticles] = useState<Particle[]>([]);
-  const [opponentParticles, setOpponentParticles] = useState<Particle[]>([]);
-
-  useEffect(() => {
-    if (myScore > prevMyScore.current) {
-      const delta = myScore - prevMyScore.current;
-      setMyFlash({ delta, key: Date.now() });
-      setMyParticles(generateParticles());
-      setTimeout(() => {
-        setMyFlash(null);
-        setMyParticles([]);
-      }, 2000);
-    }
-    prevMyScore.current = myScore;
-  }, [myScore]);
-
-  useEffect(() => {
-    const opponentScore = opponent?.score ?? 0;
-    if (opponentScore > prevOpponentScore.current) {
-      const delta = opponentScore - prevOpponentScore.current;
-      setOpponentFlash({ delta, key: Date.now() });
-      setOpponentParticles(generateParticles());
-      setTimeout(() => {
-        setOpponentFlash(null);
-        setOpponentParticles([]);
-      }, 2000);
-    }
-    prevOpponentScore.current = opponentScore;
-  }, [opponent?.score]);
-
-  const pythonStarter = `def solve(x):
+const pythonStarter = `def solve(x):
     # Write your solution here
     return None`;
 
-  const javaStarter = `public class Main {
+const javaStarter = `public class Main {
     public static void main(String[] args) {
     }
 }`;
 
+const EFFECT_DISPLAY: Record<string, { label: string; emoji: string }> = {
+  SQUID_INK_SABOTAGE: { label: "Squid Ink", emoji: "🦑" },
+  JITTER_SABOTAGE:    { label: "Earthquake", emoji: "⚡" },
+  ROTATE_SABOTAGE:    { label: "Flip Screen", emoji: "🌀" },
+};
+
+export default function GamePage() {
+  const { value: userId } = useLocalStorage("userid", "");
+  const { value: token } = useLocalStorage("token", "");
+  const { value: storedUsername } = useLocalStorage("username", "Player One");
+  const { value: storedAvatarId } = useLocalStorage("avatarId", "1");
+  const params = useParams();
+  const gameSessionId = params.gameSessionId as string;
+
+  const [language, setLanguage] = useState("python");
+  const [playerSessionId, setPlayerSessionId] = useState<number | null>(null);
+  const [players, setPlayers] = useState<Record<string, { username: string; score: number }>>({});
+  const [playerAvatarId, setPlayerAvatarId] = useState<number>(1);
+  const [opponentAvatarId, setOpponentAvatarId] = useState<number>(2);
+  const [problem, setProblem] = useState<Problem | null>(null);
+  const [currentRound, setCurrentRound] = useState(1);
+  const [gameEndTime, setGameEndTime] = useState<number | null>(null);
+  const [showSubmitToast, setShowSubmitToast] = useState(false);
   const [code, setCode] = useState(pythonStarter);
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [isRunning, setIsRunning] = useState(false);
   const [runResult, setRunResult] = useState<ExecutionResult | null>(null);
   const [submitResult, setSubmitResult] = useState<ExecutionResult | null>(null);
+  const [coinBalance, setCoinBalance] = useState<number>(0);
 
-  const toggleSolution = (id: string) => {
-    setExpandedSolutions(prev => {
-      const next = new Set(prev);
-      if (next.has(id)) next.delete(id);
-      else next.add(id);
-      return next;
-    });
+  const me = players[String(userId)];
+  const allPlayers = Object.entries(players);
+  const opponentEntry = userId ? allPlayers.find(([id]) => id !== String(userId)) : null;
+  const opponent = opponentEntry ? opponentEntry[1] : null;
+  const myScore = me?.score ?? 0;
+
+  const { isGameOver, gameEndData, gameSummary, activeEffects, sabotageNotification } = useGameWebSocket(
+    gameSessionId, token, userId, playerSessionId, setPlayers,
+  );
+  const timeLeft = useGameTimer(gameEndTime);
+  const myAnimation = useScoreAnimation(myScore);
+  const opponentAnimation = useScoreAnimation(opponent?.score ?? 0);
+
+  const fetchCoinBalance = async () => {
+    if (!userId || !token) return;
+    try {
+      const res = await fetch(`${getApiDomain()}/users/${userId}`, { headers: { token } });
+      const data = await res.json();
+      if (data.coins !== undefined) setCoinBalance(data.coins);
+    } catch (err) {
+      console.error("Failed to fetch coin balance:", err);
+    }
   };
+
+  useEffect(() => {
+    fetchCoinBalance();
+  }, [userId, token]);
 
   // Load problem data from localStorage (saved by lobby page on game-start WS)
   useEffect(() => {
@@ -242,9 +97,7 @@ export default function GamePage() {
       const initialPlayers: Record<string, { username: string; score: number }> = {
         [String(data.playerId)]: { username: storedUsername, score: data.currentScore ?? 0 },
       };
-      if (data.opponentName) {
-        initialPlayers["opponent"] = { username: data.opponentName, score: 0 };
-      }
+      if (data.opponentName) initialPlayers["opponent"] = { username: data.opponentName, score: 0 };
       setPlayers(initialPlayers);
       const lang = (data.gameLanguage ?? "python").toLowerCase();
       setLanguage(lang);
@@ -259,78 +112,11 @@ export default function GamePage() {
       });
       setPlayerAvatarId(data.playerAvatarId ?? 1);
       setOpponentAvatarId(data.opponentAvatarId ?? 2);
-
-      // Set up game end time from backend endsAt, or fall back to 15min from now
-      if (data.endsAt) {
-        setGameEndTime(new Date(data.endsAt).getTime());
-      } else {
-        setGameEndTime(Date.now() + GAME_DURATION_MS);
-      }
+      setGameEndTime(data.endsAt ? new Date(data.endsAt).getTime() : Date.now() + GAME_DURATION_MS);
     } catch (e) {
       console.error("Failed to parse game data from localStorage", e);
     }
   }, [gameSessionId, storedUsername]);
-
-  // Countdown tick — runs independently of problem changes
-  useEffect(() => {
-    if (gameEndTime === null) return;
-    const tick = () => {
-      const remaining = Math.max(0, Math.floor((gameEndTime - Date.now()) / 1000));
-      setTimeLeft(remaining);
-    };
-    tick();
-    const interval = setInterval(tick, 1000);
-    return () => clearInterval(interval);
-  }, [gameEndTime]);
-
-  useEffect(() => {
-    if (typeof window === "undefined" || !token || !gameSessionId || !userId || userId === "") return;
-
-    const client = new Client({
-      webSocketFactory: () => new SockJS(`${getApiDomain()}/ws`),
-      connectHeaders: { token: token },
-      onConnect: () => {
-        console.log("Connected to Game WebSockets as user:", userId);
-
-        client.subscribe(`/topic/game/${gameSessionId}/points-update`, (message: IMessage) => {
-          const data = JSON.parse(message.body);
-          const incomingSessionId = Number(data.playerSessionId);
-          if (incomingSessionId === playerSessionId) {
-            setPlayers(prev => ({
-              ...prev,
-              [String(userId)]: { ...prev[String(userId)], score: data.currentScore },
-            }));
-          } else {
-            setPlayers(prev => {
-              const entry = Object.entries(prev).find(([id]) => id !== String(userId));
-              const opponentId = entry ? entry[0] : "opponent";
-              return { ...prev, [opponentId]: { ...prev[opponentId], score: data.currentScore } };
-            });
-          }
-        });
-
-        client.subscribe(`/topic/game/${gameSessionId}/end`, (message: IMessage) => {
-          const endData: GameEndDTO = JSON.parse(message.body);
-          setGameEndData(endData);
-          setIsGameOver(true);
-          localStorage.removeItem("gameRoundData");
-          localStorage.removeItem("roomLanguage");
-        });
-
-        client.subscribe(`/user/queue/game-summary`, (message: IMessage) => {
-          const summary: PlayerGameSummaryDTO = JSON.parse(message.body);
-          setGameSummary(summary);
-        });
-      },
-      onStompError: (frame) => {
-        console.error("Broker reported error: " + frame.headers["message"]);
-        console.error("Additional details: " + frame.body);
-      },
-    });
-
-    client.activate();
-    return () => { if (client.active) client.deactivate(); };
-  }, [token, gameSessionId, userId]);
 
   // RUN BUTTON LOGIC
   const handleRun = async () => {
@@ -394,8 +180,8 @@ export default function GamePage() {
         }));
         const lang = (updatedGameRound.gameLanguage ?? language).toLowerCase();
         setCode(lang === "java" ? javaStarter : pythonStarter);
+        await fetchCoinBalance();
       } else if (response.status === 204) {
-        setIsGameOver(true);
         console.log("Game over or no new content.");
       }
     } catch (error) {
@@ -433,190 +219,41 @@ export default function GamePage() {
     }
   };
 
+  // BUY ITEM
+  const handleBuyItem = async (itemId: string, enumValue: string): Promise<boolean> => {
+    if (!token || playerSessionId == null) return false;
+    try {
+      const response = await fetch(`${getApiDomain()}/games/${gameSessionId}/sabotage`, {
+        method: "POST",
+        headers: { "Content-Type": "application/json", token },
+        body: JSON.stringify({ playerSessionId, item: enumValue }),
+      });
+      if (!response.ok) {
+        if (response.status === 409) console.warn("Opponent is already sabotaged");
+        if (response.status === 400) console.warn("Not enough coins");
+        return false;
+      }
+      setCoinBalance(prev => prev - 5);
+      return true;
+    } catch (error) {
+      console.error("Failed to purchase sabotage:", error);
+      return false;
+    }
+  };
+
   // GAME OVER screen
   if (isGameOver) {
-    const solvedCorrectly = gameSummary?.problemResults?.solvedCorrectly ?? [];
-    const notSolved = gameSummary?.problemResults?.notSolvedFullyCorrectly ?? [];
-
     return (
-      <div className={resultStyles.pageBackground}>
-        <div className={styles.topRow}>
-          <div className={styles.logoArea}>
-            <CodosseumLogo size={100} />
-            <div className={styles.logoTexts}>
-              <h1 className={styles.logoTitle}>Codosseum</h1>
-              <p className={styles.logoSubtitle}>Game Results</p>
-            </div>
-          </div>
-          <div className={resultStyles.headerButtons} style={{ display: "flex", gap: "12px", flexShrink: 0, whiteSpace: "nowrap" }}>
-            <button className={resultStyles.secondaryButton} onClick={() => router.push("/menu")} style={{ minWidth: "fit-content" }}>
-              Back to Menu
-            </button>
-            <button className={resultStyles.primaryButton} onClick={() => router.push("/leaderboard")} style={{ minWidth: "fit-content" }}>
-              View Leaderboard
-            </button>
-          </div>
-        </div>
-
-        <div className={resultStyles.resultsContent}>
-          <div className={resultStyles.victoryBanner}>
-            <TrophyOutlined className={resultStyles.trophyIcon} style={{ fontSize: "48px", marginBottom: "10px" }} />
-            <h1 className={resultStyles.victoryTitle}>
-              {myScore > (opponent?.score ?? 0) ? "Victory!" : myScore === opponent?.score ? "It's a Tie!" : "Defeat!"}
-            </h1>
-            <p>
-              {myScore > (opponent?.score ?? 0)
-                ? `${storedUsername} wins the battle!`
-                : myScore === opponent?.score
-                ? "Great minds think alike!"
-                : `${opponent?.username} takes the win!`}
-            </p>
-            <span className={resultStyles.sessionText}>Session {gameSessionId}</span>
-          </div>
-
-          <div className={resultStyles.playerScoreBox} style={{ display: "flex", gap: "20px", justifyContent: "center" }}>
-            <div className={`${resultStyles.playerCard} ${myScore >= (opponent?.score ?? 0) ? resultStyles.winnerCard : ""}`}>
-              <div className={resultStyles.cardHeader}>
-                <strong>{storedUsername} (You)</strong>
-                {myScore >= (opponent?.score ?? 0) && <TrophyOutlined style={{ color: "#eab308", fontSize: "24px" }} />}
-              </div>
-              <div className={resultStyles.pointsText}>
-                {myScore} <span className={resultStyles.pointsLabel}>points</span>
-              </div>
-            </div>
-            <div className={`${resultStyles.playerCard} ${opponent ? (opponent.score >= myScore ? resultStyles.winnerCard : "") : ""}`}>
-              <div className={resultStyles.cardHeader}>
-                <strong>{opponent ? opponent.username : "Opponent"}</strong>
-                {opponent && opponent.score >= myScore && <TrophyOutlined style={{ color: "#eab308", fontSize: "24px" }} />}
-              </div>
-              <div className={resultStyles.pointsText}>
-                {opponent ? opponent.score : 0} <span className={resultStyles.pointsLabel}>points</span>
-              </div>
-            </div>
-          </div>
-
-          {/* SAMPLE SOLUTIONS */}
-          {gameEndData?.gameSessionSampleSolutions &&
-            Object.keys(gameEndData.gameSessionSampleSolutions).length > 0 && (
-            <div className={resultStyles.problemsSection}>
-              <h2 style={{ fontSize: "18px", fontWeight: 700, color: "#1a1a2e", margin: "0 0 20px 0" }}>
-                Sample Solutions
-              </h2>
-              {Object.entries(gameEndData.gameSessionSampleSolutions).map(([problemId, solution], index) => {
-                const isExpanded = expandedSolutions.has(problemId);
-                const pid = Number(problemId);
-                const isSolved = solvedCorrectly.includes(pid);
-                const isIncorrect = !isSolved && notSolved.includes(pid);
-                const borderColor = isSolved ? "#16a34a" : isIncorrect ? "#dc2626" : "#d1d5db";
-                const bgColor = isSolved ? "#f0fdf4" : isIncorrect ? "#fef2f2" : "#f9fafb";
-                const badge = isSolved
-                  ? <span style={{ fontSize: 13, fontWeight: 600, color: "#16a34a", background: "#dcfce7", padding: "2px 10px", borderRadius: 6 }}>✓ Correct</span>
-                  : isIncorrect
-                  ? <span style={{ fontSize: 13, fontWeight: 600, color: "#dc2626", background: "#fee2e2", padding: "2px 10px", borderRadius: 6 }}>✗ Incorrect</span>
-                  : <span style={{ fontSize: 13, fontWeight: 600, color: "#6b7280", background: "#f3f4f6", padding: "2px 10px", borderRadius: 6 }}>— Not solved</span>;
-
-                return (
-                  <div
-                    key={problemId}
-                    className={resultStyles.problemItem}
-                    style={{ flexDirection: "column", alignItems: "flex-start", gap: "12px", cursor: "default", border: `2px solid ${borderColor}`, background: bgColor }}
-                  >
-                    <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center", width: "100%" }}>
-                      <div style={{ display: "flex", alignItems: "center", gap: "12px" }}>
-                        <div className={resultStyles.problemIndex}>{index + 1}</div>
-                        <strong style={{ fontSize: "15px", color: "#1a1a2e" }}>{solution.problemTitle}</strong>
-                        {badge}
-                      </div>
-                      <button className={resultStyles.solutionToggleBtn} onClick={() => toggleSolution(problemId)}>
-                        {isExpanded ? "Hide Solution" : "Show Solution"}
-                      </button>
-                    </div>
-                    {isExpanded && (
-                      <pre className={resultStyles.solutionCode} style={{ width: "100%", margin: 0, overflowX: "auto", boxSizing: "border-box" }}>
-                        {solution.problemSampleSolution}
-                      </pre>
-                    )}
-                  </div>
-                );
-              })}
-            </div>
-          )}
-        </div>
-      </div>
+      <GameOverScreen
+        storedUsername={storedUsername}
+        myScore={myScore}
+        opponent={opponent}
+        gameEndData={gameEndData}
+        gameSummary={gameSummary}
+        gameSessionId={gameSessionId}
+      />
     );
   }
-
-  const currentResult = runResult ?? submitResult;
-  const testCases = currentResult && "testCases" in currentResult ? currentResult.testCases : null;
-
-  // Timer colour: red under 60s, orange under 5min, default otherwise
-  const timerColor =
-    timeLeft !== null && timeLeft <= 60 ? "#dc2626"
-    : timeLeft !== null && timeLeft <= 300 ? "#d97706"
-    : undefined;
-
-  // Reusable score box renderer
-  const renderScoreBox = (
-    isMe: boolean,
-    flash: { delta: number; key: number } | null,
-    particles: Particle[],
-  ) => {
-    const boxClass = isMe
-      ? `${styles.nameBox} ${styles.nameBoxYou} ${flash ? styles.boxStrike : ""}`
-      : `${styles.nameBox} ${flash ? styles.boxStrike : ""}`;
-    const borderStyle = isMe ? {} : { border: "2px solid #ef4444" };
-    const avatarId = isMe ? Number(storedAvatarId) : opponentAvatarId;
-    const avatarBg = isMe ? undefined : "#ef4444";
-    const label = isMe ? "You" : "Opponent";
-    const name = isMe ? storedUsername : (opponent ? opponent.username : "Waiting...");
-    const score = isMe ? myScore : (opponent ? opponent.score : 0);
-    const trophyColor = isMe ? "#3b82f6" : "#ef4444";
-
-    return (
-      <div style={{ position: "relative" }}>
-        {/* Score box — overflow:hidden clips the slash */}
-        <div
-          className={boxClass}
-          style={{ display: "flex", alignItems: "center", position: "relative", overflow: "hidden", ...borderStyle }}
-        >
-          <CodosseumAvatar id={avatarId} size={50} backgroundColor={avatarBg} />
-          <div className={styles.sessionArea}>
-            <p className={styles.sessionLabel}>{label}</p>
-            <h2 className={styles.sessionValue}>{name}</h2>
-          </div>
-          <div style={{ display: "flex", alignItems: "center", gap: "4px", marginLeft: "10px" }}>
-            <TrophyOutlined style={{ color: trophyColor, fontSize: "20px" }} />
-            <span style={{ fontWeight: "700", fontSize: "22px" }}>{score}</span>
-          </div>
-          {/* Sword slash sweeps through the box */}
-          {flash && <div key={flash.key} className={styles.slashElement} />}
-        </div>
-
-        {/* Particles — outside so they aren't clipped by overflow:hidden */}
-        {particles.map(p => (
-          <div
-            key={p.id}
-            className={styles.particle}
-            style={{
-              width: p.size,
-              height: p.size,
-              background: p.color,
-              animationDuration: `${p.duration}s`,
-              "--tx": `${p.x}px`,
-              "--ty": `${p.y}px`,
-            } as React.CSSProperties}
-          />
-        ))}
-
-        {/* Floating badge — outside so it floats above without being clipped */}
-        {flash && (
-          <div key={`b-${flash.key}`} className={styles.scoreBadge}>
-            ⚔️ +{flash.delta} pts
-          </div>
-        )}
-      </div>
-    );
-  };
 
   return (
     <div className={styles.pageBackground}>
@@ -624,7 +261,7 @@ export default function GamePage() {
       {/* SUBMIT TOAST */}
       {showSubmitToast && (
         <div style={{
-          position: "fixed", top: "24px", right: "24px", zIndex: 9999,
+          position: "fixed", top: "100px", right: "24px", zIndex: 9999,
           display: "flex", alignItems: "center", gap: "10px",
           background: "#fff", border: "1.5px solid #16a34a",
           borderRadius: "10px", padding: "12px 20px",
@@ -637,6 +274,28 @@ export default function GamePage() {
           </div>
         </div>
       )}
+
+      {/* SABOTAGE NOTIFICATION */}
+      {sabotageNotification && (() => {
+        const effect = EFFECT_DISPLAY[sabotageNotification];
+        return (
+          <div style={{
+            position: "fixed", top: "100px", left: "24px", zIndex: 9999,
+            display: "flex", alignItems: "center", gap: "10px",
+            background: "#fff", border: "1.5px solid #7c3aed",
+            borderRadius: "10px", padding: "12px 20px",
+            boxShadow: "0 4px 20px rgba(0,0,0,0.12)",
+          }}>
+            <span style={{ fontSize: "20px" }}>{effect?.emoji}</span>
+            <div>
+              <div style={{ fontWeight: 700, fontSize: "14px", color: "#7c3aed" }}>
+                {opponent?.username ?? "Opponent"} used {effect?.label}!
+              </div>
+              <div style={{ fontSize: "12px", color: "#6b7280" }}>Effect active for 10 seconds</div>
+            </div>
+          </div>
+        );
+      })()}
 
       {/* HEADER */}
       <div className={styles.topRow}>
@@ -662,153 +321,56 @@ export default function GamePage() {
           </div>
 
           <div style={{ display: "flex", gap: "20px", alignItems: "center", justifyContent: "flex-end", width: "100%", paddingRight: "30px" }}>
-            {renderScoreBox(true, myFlash, myParticles)}
+            <ScoreBox
+              isMe={true}
+              storedUsername={storedUsername}
+              storedAvatarId={Number(storedAvatarId)}
+              opponentAvatarId={opponentAvatarId}
+              opponentUsername={opponent?.username ?? null}
+              myScore={myScore}
+              opponentScore={opponent?.score ?? 0}
+              flash={myAnimation.flash}
+              particles={myAnimation.particles}
+            />
             <span style={{ fontWeight: "bold", color: "#94a3b8" }}>VS</span>
-            {renderScoreBox(false, opponentFlash, opponentParticles)}
+            <ScoreBox
+              isMe={false}
+              storedUsername={storedUsername}
+              storedAvatarId={Number(storedAvatarId)}
+              opponentAvatarId={opponentAvatarId}
+              opponentUsername={opponent?.username ?? null}
+              myScore={myScore}
+              opponentScore={opponent?.score ?? 0}
+              flash={opponentAnimation.flash}
+              particles={opponentAnimation.particles}
+            />
           </div>
         </div>
       </div>
 
       {/* MAIN CONTENT */}
-      <div className={styles.content}>
+      <div
+        className={`${styles.content} ${activeEffects.has("jitter") ? styles.jitterActive : ""} ${activeEffects.has("rotate") ? styles.rotateActive : ""}`}
+      >
+        {activeEffects.has("ink") && <div className={styles.inkOverlay} />}
 
-        {/* LEFT: PROBLEM */}
-        <div className={styles.card}>
-          {problem ? (
-            <>
-              <section className={styles.section}>
-                <div className={styles.problemHeader}>
-                  <h3 className={styles.problemTitle}>{problem.title}</h3>
-                  <div className={styles.badgeRow}>
-                    <span className={styles.languageIndicator}>
-                      {language.charAt(0).toUpperCase() + language.slice(1)}
-                    </span>
-                    <span
-                      className={styles.timerBadge}
-                      style={timerColor ? { color: timerColor, borderColor: timerColor } : undefined}
-                    >
-                      <ClockCircleOutlined />
-                      {timeLeft !== null ? formatTime(timeLeft) : "15:00"}
-                    </span>
-                  </div>
-                </div>
-              </section>
-
-              <hr className={styles.divider} />
-
-              <div className={styles.scrollableContent}>
-                <section className={styles.section}>
-                  <h3 className={styles.sectionTitle}>Description</h3>
-                  <p className={styles.problemDescription}>{problem.description}</p>
-                </section>
-                <section className={styles.section}>
-                  <h3 className={styles.sectionTitle}>Input Format</h3>
-                  <div className={styles.exampleCard}>
-                    <p className={styles.exampleText}>{problem.inputFormat}</p>
-                  </div>
-                </section>
-                <section className={styles.section}>
-                  <h3 className={styles.sectionTitle}>Output Format</h3>
-                  <div className={styles.exampleCard}>
-                    <p className={styles.exampleText}>{problem.outputFormat}</p>
-                  </div>
-                </section>
-                <section className={styles.section}>
-                  <h3 className={styles.sectionTitle}>Constraints</h3>
-                  <div className={styles.exampleCard}>
-                    <p className={styles.exampleText}>{problem.constraints}</p>
-                  </div>
-                </section>
-              </div>
-            </>
-          ) : (
-            <div style={{ display: "flex", alignItems: "center", justifyContent: "center", flex: 1 }}>
-              <p style={{ color: "#6b7280" }}>Loading problem...</p>
-            </div>
-          )}
+        {/* LEFT COLUMN */}
+        <div style={{ display: "flex", flexDirection: "column", gap: "16px", flex: 1, minWidth: 0 }}>
+          <ProblemPanel problem={problem} language={language} timeLeft={timeLeft} />
+          <ItemShop coinBalance={coinBalance} onBuyItem={handleBuyItem} />
         </div>
 
-        {/* RIGHT: CODE EDITOR */}
-        <div style={{ display: "flex", flexDirection: "column", gap: "16px", flex: 1 }}>
-          <div className={styles.card} style={{ flex: 2, paddingBottom: 0, overflow: "hidden" }}>
-            <section className={styles.section}>
-              <div className={styles.problemHeader}>
-                <h3 className={styles.sectionTitle}>Code Editor</h3>
-                <span className={styles.languageIndicator}>
-                  {language.charAt(0).toUpperCase() + language.slice(1)}
-                </span>
-              </div>
-            </section>
-            <hr className={styles.divider} />
-            <div className={styles.editorBox}>
-              <CodeMirror
-                value={code}
-                height="100%"
-                style={{ height: "100%" }}
-                extensions={[language === "java" ? java() : python(), indentUnit.of("    ")]}
-                onChange={(value) => setCode(value)}
-                basicSetup={{ lineNumbers: true, foldGutter: false, dropCursor: true, allowMultipleSelections: true, indentOnInput: true }}
-              />
-            </div>
-            <div className={styles.actionRow}>
-              <button className={styles.runButton} onClick={handleRun} disabled={isRunning || isSubmitting}>
-                {isRunning ? <LoadingOutlined spin /> : <><PlayCircleOutlined /> Run</>}
-              </button>
-              <button className={styles.submitButton} onClick={handleSubmit} disabled={isRunning || isSubmitting}>
-                {isSubmitting ? <LoadingOutlined spin /> : <><SendOutlined /> Submit</>}
-              </button>
-            </div>
-          </div>
-
-          {/* OUTPUT RENDERING */}
-          <div className={styles.card} style={{ flex: 1 }}>
-            <section className={styles.section}>
-              <h3 className={styles.sectionTitle}>
-                <svg className={styles.outputTitleIcon} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round">
-                  <polyline points="4 17 10 11 4 5"></polyline>
-                  <line x1="12" y1="19" x2="20" y2="19"></line>
-                </svg>
-                Output
-              </h3>
-            </section>
-            <hr className={styles.divider} />
-            <div className={styles.outputContent}>
-              {testCases ? (
-                <div style={{ display: "flex", flexDirection: "column", gap: "12px", padding: "0 8px 20px 8px" }}>
-                  {currentResult?.summary && (
-                    <div style={{ fontWeight: 600, fontSize: "16px", color: currentResult.status === "success" ? "#16a34a" : "#dc2626" }}>
-                      {currentResult.summary}
-                    </div>
-                  )}
-                  {testCases.map((t, index) => {
-                    const isPass = t.result === "PASS";
-                    return (
-                      <div key={t.testCaseId} style={{ border: `1px solid ${isPass ? "#16a34a" : "#dc2626"}`, borderRadius: "8px", padding: "10px", background: isPass ? "#f0fdf4" : "#fef2f2" }}>
-                        <div style={{ fontWeight: 600 }}>{isPass ? "✅ PASS" : "❌ FAIL"} — Test {index + 1}</div>
-                        <div style={{ marginTop: "6px", fontSize: "13px" }}>
-                          <div><strong>Expected:</strong> {t.expectedOutput}</div>
-                          <div><strong>Actual:</strong> {t.actualOutput}</div>
-                        </div>
-                      </div>
-                    );
-                  })}
-                </div>
-              ) : currentResult?.message ? (
-                <pre className={styles.exampleText} style={{ padding: "20px", whiteSpace: "pre-wrap" }}>
-                  {currentResult.message}
-                </pre>
-              ) : (
-                <div className={styles.placeholderContainer}>
-                  <svg className={styles.terminalIcon} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round">
-                    <polyline points="4 17 10 11 4 5"></polyline>
-                    <line x1="12" y1="19" x2="20" y2="19"></line>
-                  </svg>
-                  <p className={styles.placeholderText}>Run or submit your code to see results</p>
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
+        <CodeEditorPanel
+          code={code}
+          setCode={setCode}
+          language={language}
+          isRunning={isRunning}
+          isSubmitting={isSubmitting}
+          onRun={handleRun}
+          onSubmit={handleSubmit}
+          runResult={runResult}
+          submitResult={submitResult}
+        />
       </div>
     </div>
   );
