+++ app/styles/game.module.css
@@ -29,6 +29,7 @@
     box-sizing: border-box;
     align-items: stretch;
     overflow: hidden;
+    position: relative;
 }
 
 .logoArea {
@@ -92,7 +93,6 @@
     display: flex;
     align-items: center;
     gap: 8px;
-
 }
 
 .problemDescription {
@@ -113,7 +113,6 @@
     display: flex;
     align-items: center;
     gap: 8px;
-
 }
 
 .editorContainer {
@@ -159,13 +158,11 @@
     line-height: 1.6;
 }
 
-
 .editorBox :global(.cm-content),
 .editorBox :global(.cm-line) {
     font-family: 'Fira Code', 'JetBrains Mono', monospace !important;
     padding: 0 4px !important;
     min-height: 100%;
-
 }
 
 .editorBox :global(.cm-gutters) {
@@ -380,7 +377,6 @@
     flex-direction: column;
     gap: 16px;
     border: 2px solid #e2e8f0;
-
 }
 
 .exampleBlock {
@@ -419,7 +415,6 @@
     overflow-y: auto;
     min-height: 100%;
     padding-bottom: 20px;
-
 }
 
 .placeholderContainer {
@@ -493,33 +488,33 @@
     8%   { opacity: 1; }
     92%  { opacity: 0.9; }
     100% { transform: translateX(420%) skewX(-15deg); opacity: 0; }
-  }
-  
-  .slashElement {
+}
+
+.slashElement {
     position: absolute;
     top: -20%;
     left: 0;
     width: 70px;
     height: 140%;
     background: linear-gradient(
-      100deg,
-      transparent 10%,
-      rgba(255, 215, 0, 0.15) 30%,
-      rgba(255, 255, 255, 0.95) 50%,
-      rgba(255, 215, 0, 0.15) 70%,
-      transparent 90%
+        100deg,
+        transparent 10%,
+        rgba(255, 215, 0, 0.15) 30%,
+        rgba(255, 255, 255, 0.95) 50%,
+        rgba(255, 215, 0, 0.15) 70%,
+        transparent 90%
     );
     animation: swordSlash 0.38s cubic-bezier(0.16, 1, 0.3, 1) forwards;
     pointer-events: none;
     z-index: 10;
-  }
-  
-  @keyframes particleFly {
+}
+
+@keyframes particleFly {
     0%   { transform: translate(-50%, -50%) translate(0px, 0px) scale(1); opacity: 1; }
     100% { transform: translate(-50%, -50%) translate(var(--tx), var(--ty)) scale(0); opacity: 0; }
-  }
-  
-  .particle {
+}
+
+.particle {
     position: absolute;
     top: 50%;
     left: 50%;
@@ -527,15 +522,15 @@
     animation: particleFly 1.1s cubic-bezier(0.25, 0.46, 0.45, 0.94) forwards;
     pointer-events: none;
     z-index: 30;
-  }
-  
-  @keyframes scoreFloat {
+}
+
+@keyframes scoreFloat {
     0%   { opacity: 1; transform: translateY(0px) scale(1); }
     18%  { opacity: 1; transform: translateY(-8px) scale(1.3); }
     100% { opacity: 0; transform: translateY(-52px) scale(0.85); }
-  }
-  
-  .scoreBadge {
+}
+
+.scoreBadge {
     position: absolute;
     top: -6px;
     right: 6px;
@@ -547,14 +542,62 @@
     color: #eab308;
     text-shadow: 0 2px 12px rgba(0, 0, 0, 0.6);
     white-space: nowrap;
-  }
-  
-  @keyframes boxStrike {
+}
+
+@keyframes boxStrike {
     0%   { box-shadow: 0 0 0 0 rgba(234, 179, 8, 0); }
     20%  { box-shadow: 0 0 18px 5px rgba(234, 179, 8, 0.75); }
     100% { box-shadow: 0 0 0 0 rgba(234, 179, 8, 0); }
-  }
-  
-  .boxStrike {
+}
+
+.boxStrike {
     animation: boxStrike 0.75s ease-out !important;
-  }
+}
+
+.inkOverlay {
+    position: absolute;
+    inset: 0;
+    pointer-events: none;
+    z-index: 9000;
+    overflow: hidden;
+}
+
+.inkOverlay::before {
+    content: '';
+    position: absolute;
+    inset: 0;
+    background: rgba(8, 2, 25, 0.97);
+    animation: inkSplash 0.5s cubic-bezier(0.22, 1, 0.36, 1) forwards;
+    transform-origin: 52% 48%;
+}
+
+
+
+
+@keyframes inkSplash {
+    from { transform: scale(0); opacity: 0; }
+    to   { transform: scale(1); opacity: 1; }
+}
+
+.jitterActive {
+    animation: jitter 0.08s linear infinite;
+}
+
+@keyframes jitter {
+    0%   { transform: translate(0, 0) rotate(0deg); }
+    10%  { transform: translate(-8px, 5px) rotate(-0.5deg); }
+    20%  { transform: translate(9px, -7px) rotate(0.4deg); }
+    30%  { transform: translate(-7px, 8px) rotate(-0.6deg); }
+    40%  { transform: translate(10px, -5px) rotate(0.5deg); }
+    50%  { transform: translate(-9px, 4px) rotate(-0.4deg); }
+    60%  { transform: translate(7px, 9px) rotate(0.6deg); }
+    70%  { transform: translate(-10px, -6px) rotate(-0.5deg); }
+    80%  { transform: translate(8px, 7px) rotate(0.4deg); }
+    90%  { transform: translate(-6px, -8px) rotate(-0.6deg); }
+    100% { transform: translate(0, 0) rotate(0deg); }
+}
+
+.rotateActive {
+    transform: rotate(180deg);
+    transition: transform 0.6s cubic-bezier(0.68, -0.55, 0.27, 1.55);
+}
