+++ app/styles/itemShop.module.css
@@ -1,0 +1,109 @@
+.shopCard {
+    background: #ffffff;
+    border: 1.5px solid #e5e7eb;
+    border-radius: 12px;
+    padding: 14px 16px;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
+    flex-shrink: 0;
+}
+
+.shopHeader {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 12px;
+}
+
+.shopTitle {
+    font-weight: 700;
+    font-size: 16px;
+    color: #1a1a2e;
+}
+
+.coinBalance {
+    display: flex;
+    align-items: center;
+    gap: 5px;
+    background: #fef9c3;
+    border: 1px solid #fde047;
+    border-radius: 20px;
+    padding: 3px 10px;
+    font-size: 14px;
+}
+
+.coinAmount {
+    font-weight: 700;
+    font-size: 14px;
+    color: #854d0e;
+}
+
+.itemsGrid {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 10px;
+}
+
+.itemCard {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 6px;
+    border: 1.5px solid;
+    border-radius: 10px;
+    padding: 16px 8px;
+    text-align: center;
+    transition: opacity 0.2s, border-color 0.2s;
+}
+
+.itemEmoji {
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 20px;
+    flex-shrink: 0;
+}
+
+.itemInfo {
+    display: flex;
+    flex-direction: column;
+    gap: 2px;
+    flex: 1;
+    width: 100%;
+}
+
+.itemName {
+    font-weight: 700;
+    font-size: 12px;
+    color: #1a1a2e;
+}
+
+.itemDesc {
+    font-size: 10px;
+    color: #6b7280;
+    line-height: 1.3;
+}
+
+.itemDuration {
+    font-size: 10px;
+    color: #9ca3af;
+    font-style: italic;
+}
+
+.buyButton {
+    width: 100%;
+    border: none;
+    border-radius: 6px;
+    padding: 5px 6px;
+    font-size: 11px;
+    font-weight: 700;
+    color: #ffffff;
+    transition: opacity 0.2s;
+    margin-top: 2px;
+}
+
+.buyButton:not(:disabled):hover {
+    opacity: 0.85;
+}
